+++ src/extraction/render_forms.py
@@ -9,7 +9,7 @@ from typing import Any
 
 from PIL import Image, ImageDraw, ImageFont
 
-from src.utils.io import load_jsonl, write_jsonl
+from src.utils.io import cache_safe_id, load_jsonl, write_jsonl
 
 
 def _font(size: int = 14) -> ImageFont.ImageFont:
@@ -179,7 +179,7 @@ def render_documents(docs_path: Path, out_dir: Path) -> Path:
     for doc in load_jsonl(docs_path):
         img, words, truncated = render_page(doc["text"])
         labeled = label_words(words)
-        img_path = images_dir / f"{doc['record_id'].replace('::', '__')}.png"
+        img_path = images_dir / f"{cache_safe_id(str(doc['record_id']))}.png"
         img.save(img_path)
         rows.append(
             {